Move error enum to new file
